requirements.txt
numpy
pandas
yolo
Pillow
